requirements.txt
pytest~=7.4.0
requests~=2.31.0
zeep~=4.2.1
paramiko~=3.3.1
PyYAML~=6.0.1